requirements.txt
# Core dependencies
fastmcp>=2.12.4
mcp>=1.16.0
httpx>=0.28.1
uvicorn[standard]==0.34.0
pydantic==2.11.10
docker==7.1.0
gitpython==3.1.45
PyYAML==6.0.3
redis==6.4.0

# Development dependencies
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-mock==3.15.1
black==25.9.0
isort==6.1.0
flake8==7.3.0
mypy==1.18.2
autopep8==2.3.2